+++ games/X-Com - Terror from the Deep/SR-Tactical/main.c
@@ -1,6 +1,6 @@
 /**
  *
- *  Copyright (C) 2016-2021 Roman Pauer
+ *  Copyright (C) 2016-2022 Roman Pauer
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy of
  *  this software and associated documentation files (the "Software"), to deal in
@@ -2004,18 +2004,18 @@ static void Game_Event_Loop(void)
                     case EC_MOUSE_MOVE:
                         {
                         #ifdef USE_SDL2
-                            SDL_WarpMouseInWindow(Game_Window, Game_MouseX + (int) event.user.data1, Game_MouseY + (int) event.user.data2);
+                            SDL_WarpMouseInWindow(Game_Window, Game_MouseX + (intptr_t) event.user.data1, Game_MouseY + (intptr_t) event.user.data2);
                         #else
-                            SDL_WarpMouse(Game_MouseX + (int) event.user.data1, Game_MouseY + (int) event.user.data2);
+                            SDL_WarpMouse(Game_MouseX + (intptr_t) event.user.data1, Game_MouseY + (intptr_t) event.user.data2);
                         #endif
                         }
                         break;
                     case EC_MOUSE_SET:
                         {
                         #ifdef USE_SDL2
-                            SDL_WarpMouseInWindow(Game_Window, (int) event.user.data1, (int) event.user.data2);
+                            SDL_WarpMouseInWindow(Game_Window, (intptr_t) event.user.data1, (intptr_t) event.user.data2);
                         #else
-                            SDL_WarpMouse((int) event.user.data1, (int) event.user.data2);
+                            SDL_WarpMouse((intptr_t) event.user.data1, (intptr_t) event.user.data2);
                         #endif
                         }
                         break;
@@ -2025,7 +2025,7 @@ static void Game_Event_Loop(void)
                         break;
                     //senquack - need this to allow macros to behave in battlescape
                     case EC_DELAY:
-                        SDL_Delay((int) event.user.data1);
+                        SDL_Delay((intptr_t) event.user.data1);
                         break;
                         // case EC_DELAY:
